exmdb_provider: resolve compiler warnings on x32
exch/exmdb_provider/common_util.cpp: In function "long int cu_set_object_cid_value(sqlite3*, db_table, uint64_t, const TAGGED_PROPVAL*)":
exch/exmdb_provider/common_util.cpp:2770:45: warning: comparison of integer expressions of different signedness: "ssize_t" {aka "int"} and "uint32_t" {aka "unsigned int"} [-Wsign-compare]
 2770 | if (write(fd.get(), bv->pv, bv->cb) != bv->cb ||
exch/exmdb_provider/notification_agent.cpp: In function "void notification_agent_thread_work(std::shared_ptr<ROUTER_CONNECTION>&&)":
exch/exmdb_provider/notification_agent.cpp:92:43: warning: comparison of integer expressions of different signedness: "int" and "uint32_t" {aka "unsigned int"} [-Wsign-compare]
   92 | if (bytes_written != dg.cb ||
exch/exmdb_provider/ics.cpp: In function "long int exmdb_server_get_content_sync(const char*, uint64_t, const char*, const IDSET*, const IDSET*, const IDSET*, const IDSET*, uint32_t, const RESTRICTION*, long int, uint32_t*, uint64_t*, uint32_t*, uint64_t*, EID_ARRAY*, EID_ARRAY*, uint64_t*, EID_ARRAY*, EID_ARRAY*, EID_ARRAY*, EID_ARRAY*, EID_ARRAY*, uint64_t*)":
exch/exmdb_provider/ics.cpp:525:57: warning: comparison of integer expressions of different signedness: "uint32_t" {aka "unsigned int"} and "ssize_t" {aka "int"} [-Wsign-compare]
  525 | if (punread_mids->count == count ||
exch/exmdb_provider/ics.cpp:526:55: warning: comparison of integer expressions of different signedness: "uint32_t" {aka "unsigned int"} and "ssize_t" {aka "int"} [-Wsign-compare]
  526 | pread_mids->count == count)
exch/exmdb_provider/folder.cpp: In function "long int exmdb_server_set_folder_by_class(const char*, uint64_t, const char*, long int*)":
exch/exmdb_provider/folder.cpp:116:43: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "uint64_t" {aka "long long unsigned int"} [-Wformat=]
  116 | " VALUES (?, ?, %lu)", rop_util_current_nttime());
exch/exmdb_provider/store.cpp:29:53: warning: unsigned conversion from "long long int" to "size_t" {aka "unsigned int"} changes value from "140737488355327" to "4294967295" [-Woverflow]
   29 | static constexpr size_t MAXIMUM_ALLOCATION_NUMBER = 0x7fffffffffffLL;
exch/exmdb_provider/store.cpp: In function "long int exmdb_server_allocate_ids(const char*, uint32_t, uint64_t*)":
exch/exmdb_provider/store.cpp:296:58: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "time_t" {aka "long long int"} [-Wformat=]
  296 | " WHERE allocate_time>=%lu",

+++ exch/exmdb_provider/ics.cpp
@@ -500,7 +500,7 @@ BOOL exmdb_server_get_content_sync(const char *dir,
 		if (stm_select_rd == nullptr ||
 		    sqlite3_step(stm_select_rd) != SQLITE_ROW)
 			return FALSE;
-		ssize_t count = sqlite3_column_int64(stm_select_rd, 0);
+		uint64_t count = sqlite3_column_int64(stm_select_rd, 0);
 		stm_select_rd.finalize();
 		pread_mids->count = 0;
 		punread_mids->count = 0;
